use separate views and endpoints for admin pages
