+++ helpme/templates/helpme/team_detail.html
@@ -47,7 +47,7 @@
 </p>
 {% endif %}
 <div class="row justify-content-center mb-2">
-  <a class="text-light" href="{% url 'helpme:team-list' %}">
+  <a class="text-light" href="{% url 'helpme_admin:team-list' %}">
     <button class="btn btn-primary font-weight-bold px-4 py-2 rounded-pill">
       Return to Teams Page
     </button>
