+++ helpme/templates/helpme/team_form.html
@@ -14,7 +14,7 @@
     {% for team in teams %}
     <tr>
       <td class="align-middle">
-	<a href="{% url 'helpme:team-detail' team.uuid %}"> {{ team }} </a>
+	<a href="{% url 'helpme_admin:team-detail' team.uuid %}"> {{ team }} </a>
       </td>
     </tr>
     {% empty %}
@@ -26,7 +26,7 @@
 </table>
 
 <h2>Create a new team</h2>
-<form action="{% url 'helpme:team-list' %}" method="POST" class="mb-2">
+<form action="{% url 'helpme_admin:team-list' %}" method="POST" class="mb-2">
     {% csrf_token %}
     {{ form|crispy }}
     <input type="submit" value="Create Team" class="btn btn-primary font-weight-bold py-2 px-4 mt-3 rounded-pill">
@@ -43,7 +43,7 @@
     {% for team in user_teams %}
     <tr>
       <td class="align-middle">
-	<a href="{% url 'helpme:team-detail' team.uuid %}"> {{ team }} </a>
+	<a href="{% url 'helpme_admin:team-detail' team.uuid %}"> {{ team }} </a>
       </td>
     </tr>
     {% empty %}
