+++ helpme/templates/helpme/ticket_detail.html
@@ -153,7 +153,7 @@
   </div>
 </div>
 <div class="row justify-content-center my-3">
-  <a class="text-light" href="{% url 'helpme:dashboard' %}">
+  <a class="text-light" href="{% url 'helpme_admin:dashboard' %}">
     <button class="btn btn-primary font-weight-bold px-4 py-2 rounded-pill">
       Return to Dashboard
     </button>
