+++ helpme/urls/helpme_admin.py
@@ -1,14 +1,12 @@
 from django.urls import path
 
-from helpme import views
+from helpme.views import helpme_admin as views
 
-app_name = "helpme"
+app_name = "helpme_admin"
 
 urlpatterns = [
-    path("", views.SupportDashboardView.as_view(), name="dashboard"),
-    path("faq/", views.FAQView.as_view(), name="faq"),
     path("faq/create/", views.FAQCreateView.as_view(), name="faq-create"),
-    path("request/", views.SupportRequestView.as_view(), name="submit-request"),
+    path("tickets/", views.AdminSupportDashboardView.as_view(), name="dashboard"),
     path("ticket/<uuid:uuid>/", views.TicketDetailView.as_view(), name="ticket-detail"),
     path("teams/", views.TeamCreateView.as_view(), name="team-list"),
     path("teams/<uuid:uuid>/", views.TeamDetailView.as_view(), name="team-detail"),
